rename a couple badly named modules
