+++ kernel/src/scan/log_replay.rs
